03_backend_developement/05_promises/d3df63ab202e290c26694e701b5122b8cd65073d/02_chuck_fetch/src/chuck.ts
import fetch, { Response } from "node-fetch";

function getCategories(): Promise<void> {
  return fetch(`https://api.chucknorris.io/jokes/categories`, {
    method: "GET",
  })
    .then((response) => {
      console.log(`${response.body}`);
    })
    .catch((error) => {
      console.warn(error);
    });
}

function getChuckNorrisJoke(category: string): Promise<void> {
  return fetch(`https://api.chucknorris.io/jokes/random?category=${category}`, {
    method: "GET",
  })
    .then((response) => {
      console.log(`${response.body}`);
    })
    .catch((error) => {
      console.warn(error);
    });
}

// Leave the line below for tests to work properly
export { getCategories, getChuckNorrisJoke };
